MERGE: remove Makefile -> cmakelist

+++ src/parser/parser.hh
@@ -1,9 +1,9 @@
 #ifndef PARSER_HH
 #define PARSER_HH
-#include "../pathtracer.hh"
+#include "../render.hh"
 
 namespace pathtracing {
-    void parse_arguments(Pathtracer& pt, int argc, char **argv, int& x, int& y);
+    void parse_arguments(Pathtracer& pt, int argc, char **argv, double& x, double& y);
 }
 
 #endif /* PARSER_HH */
